requirements.txt
python-telegram-bot==13
requests~=2.28.1
python-dotenv~=0.21.0